+++ src/globals/plugin-setting.ts
@@ -5,7 +5,9 @@ import { isDevEnvironment } from "src/utils/my-lib";
 export class PluginSetting {
 	mySetting: string;
 	logLevel: LogLevel;
-	enableCjkPatch: boolean;
+	enableStopWordsEn: boolean;
+	enableChinesePatch: boolean;
+	enableStopWordsZh: boolean;
 	apiProvider1: ApiProvider;
 	apiProvider2: ApiProvider;
 	excludeExtensions: string[];
@@ -16,7 +18,9 @@ export class PluginSetting {
 export const DEFAULT_PLUGIN_SETTING: PluginSetting = {
 	mySetting: "default",
 	logLevel: isDevEnvironment ? "debug" : "none",
-	enableCjkPatch: false,
+	enableStopWordsEn: false,
+	enableChinesePatch: false,
+	enableStopWordsZh: false,
 	apiProvider1: {
 		domain: "",
 		key: "",